tests/wpt/tests/css/css-view-transitions/scoped/element-active-view-transition.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>View Transition: element.activeViewTransition attribute</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-1/#dom-document-activeviewtransition">
<link rel="author" href="mailto:vmpstr@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>

<style>
#target {
  width: 100px;
  height: 100px;
}
::view-transition-group(*) {
  animation-duration: 1ms;
}
</style>

<div id=target></div>

<script>
promise_test(async t => {
  assert_implements(target.startViewTransition, "View Transitions are not supported");

  assert_equals(target.activeViewTransition, null, "activeViewTransition is null initially");

  const transition = target.startViewTransition(() => {});
  assert_equals(target.activeViewTransition, transition, "activeViewTransition returns the running transition");
  assert_equals(document.activeViewTransition, null, "doc.activeViewTransition returns null");

  await transition.finished;

  assert_equals(target.activeViewTransition, null, "activeViewTransition is null after transition finishes");


  const transition2 = target.startViewTransition(() => {});
  assert_equals(target.activeViewTransition, transition2, "activeViewTransition returns the running transition (2)");

  transition2.skipTransition();

  assert_equals(target.activeViewTransition, null, "activeViewTransition is null after transition is skipped");

  let [result] = await Promise.allSettled([transition2.ready]);
  assert_equals(result.status, "rejected", "ready promise should be rejected due to skipTransition call");
}, "document.activeViewTransition returns the active transition");
</script>
